+++ docs/changelog.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add support for Python 3.14, PyPy 3.11
 
+### Removed
+- Remove support for Python 3.8, 3.9 (includes PyPy versions).
+
 ## [4.0.0](https://github.com/nhairs/python-json-logger/compare/v3.3.3...v4.0.0) - 2025-10-06
 
 ### Added